Create pod directly
The peer is still managing the chaincode lifecycle so a deployment probably doesn’t make sense
